refactor: added global actions and simplification

+++ src/scripts/tests/sqpc/spring-normal/test.sh
@@ -9,16 +9,18 @@ TESTID=${2:-}
 
 case "${ACTION}" in
     app_start)
-        start_app "${NAME}" "${TEST_BUILDS_DIR}/${REPO_NAME}/springboot3/spring-normal/application/springboot3.jar"
+        start_app "${TESTID}" "${TEST_TEST_CACHE}/repo/springboot3/target/application/springboot3.jar"
         ;;
     setup)
+        REPO_URL="https://github.com/quarkusio/spring-quarkus-perf-comparison.git"
+        clone "${REPO_URL}"
+        [[ $CLONE_CHANGED -eq 1 ]] || return 0
         # Compile Spring Boot app normally
         require_java "21+"
-        compile_maven "${REPO_NAME}/springboot3" ""
+        compile_maven "repo/springboot3" ""
         echo "   - Extracting Spring Boot Buildpack Executable..."
-        target="${TEST_APPS_DIR}/${REPO_NAME}/springboot3/target"
+        target="${TEST_TEST_CACHE}/repo/springboot3/target"
         rm -rf "${target}/application" > /dev/null 2>&1
         java -Djarmode=tools -jar "${target}/springboot3.jar" extract --destination "${target}/application" > /dev/null
-        copy_build_artifacts "${REPO_NAME}/springboot3" "spring-normal" "target/application"
         ;;
 esac
