+++ src/scripts/tests/sqpc/spring-sbaot/test.sh
@@ -9,19 +9,21 @@ TESTID=${2:-}
 
 case "${ACTION}" in
     app_start)
-        TEST_JAVA_OPTS="${TEST_JAVA_OPTS:-} -Dspring.aot.enabled=true" start_app "${TESTID}" "${TEST_BUILDS_DIR}/${REPO_NAME}/springboot3/spring-buildpack/application/springboot3.jar"
+        TEST_JAVA_OPTS="${TEST_JAVA_OPTS:-} -Dspring.aot.enabled=true" start_app "${TESTID}" "${TEST_TEST_CACHE}/repo/springboot3/target/application/springboot3.jar"
         ;;
     setup)
+        REPO_URL="https://github.com/quarkusio/spring-quarkus-perf-comparison.git"
+        clone "${REPO_URL}"
+        [[ $CLONE_CHANGED -eq 1 ]] || return 0
         # Compile Spring Boot app as Spring Boot Buildpack Executable
         # Which means preparing for AOT cache and production environment
         # As described in https://docs.spring.io/spring-boot/reference/packaging/efficient.html
         # and in https://docs.spring.io/spring-boot/reference/packaging/aot.html
         require_java "21+"
-        compile_maven "${REPO_NAME}/springboot3" "-Pnative"
+        compile_maven "repo/springboot3" "-Pnative"
         echo "   - Extracting Spring Boot Buildpack Executable..."
-        target="${TEST_APPS_DIR}/${REPO_NAME}/springboot3/target"
+        target="${TEST_TEST_CACHE}/repo/springboot3/target"
         rm -rf "${target}/application" > /dev/null 2>&1
         java -Djarmode=tools -jar "${target}/springboot3.jar" extract --destination "${target}/application" > /dev/null
-        copy_build_artifacts "${REPO_NAME}/springboot3" "spring-buildpack" "target/application"
         ;;
 esac
